+++ test/routes/api_routes_test.rb
@@ -11,4 +11,7 @@ class ApiRoutesTest < ActionController::TestCase
   test "GET /users/:id" do
   	assert_routing({ method: 'get', path: '/users/1'}, { controller: 'users', action: 'show', id: "1" })
   end
+  test "GET /users/nearby/" do
+    assert_routing({ method: 'get', path: '/users/nearby'}, { controller: 'users', action: 'nearby' })
+  end
 end