+++ CHANGELOG.md
@@ -1,6 +1,5 @@
 ## 0.1.3 / Unrealesed
 ### API
-- show the nearest users
 - add user to bookmarks
 - show bookmarked users
 - remove user from bookmarks
@@ -8,6 +7,7 @@
 ## 0.1.2 / 15.03.2015
 ### API
 - show ad of the nearest place from user
+- search nearby users
 
 ## 0.1.1 / 14.03.2015
 ### API